Pass: 문자열 다루기 기본 Swift

+++ Programmers/Level1/Swift/문자열_다루기_기본.swift
@@ -1,6 +1,6 @@
 func solution(_ s:String) -> Bool {
     
-    if (s.count == 4 || s.count == 6) && (Double(s) != nil){
+    if (s.count == 4 || s.count == 6) && (Int(s) != nil){
         return true
     }
